# projects/windlabx4/contracts/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: WindLabX4 API
  version: "0.1.0-rebuild"
  description: |
    MVP API for the rebuilt WindLabX4 backend.
    This contract intentionally documents only endpoints currently implemented
    by projects/WindLabX4/services/api-go/api/server.go.
servers:
  - url: http://localhost:8080
    description: Local development API server
paths:
...
  /api/device/scan:
    get:
      summary: Scan for discoverable devices
      description: Returns discovered devices. MVP wiring always includes a simulated scanner so no-hardware mode remains testable.
      operationId: scanDevices
      tags:
        - Device
      responses:
        "200":
          description: Scan results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeviceScanResult"
        "400":
          description: Scan failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
    DeviceType:
      type: string
      enum:
        - SIMULATED
        - DAQ_T_1603
...
    DeviceScanResult:
      type: object
      required:
        - id
        - name
        - type
        - available
      properties:
        id:
          type: string
          example: sim-1
        name:
          type: string
          example: Simulator 1
        type:
          $ref: "#/components/schemas/DeviceType"
        available:
          type: boolean
          example: true
        address:
          type: string
          example: simulated://sim-1
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: not found